게시글 및 댓글기능 token 유저 인증방식 추가

+++ src/main/java/io/oopy/coding/api/comment/service/CommentService.java
@@ -1,5 +1,11 @@
 package io.oopy.coding.api.comment.service;
 
+import io.oopy.coding.api.comment.exception.CommentErrorCode;
+import io.oopy.coding.api.comment.exception.CommentErrorException;
+import io.oopy.coding.api.content.exception.ContentErrorCode;
+import io.oopy.coding.api.content.exception.ContentErrorException;
+import io.oopy.coding.common.response.SuccessResponse;
+import io.oopy.coding.common.security.authentication.CustomUserDetails;
 import io.oopy.coding.domain.comment.dto.*;
 import io.oopy.coding.domain.comment.entity.Comment;
 import io.oopy.coding.domain.comment.repository.CommentRepository;
@@ -27,11 +33,12 @@ public class CommentService {
      * @param contentId
      */
     @Transactional
-    public List<CommentDTO> getComments(Long contentId) {
-        Content content = contentRepository.findById(contentId)
-                .orElseThrow(() -> new EntityNotFoundException("Content does not exist"));
+    public SuccessResponse getComments(Long contentId) {
+        contentRepository.findById(contentId)
+                .orElseThrow(() -> new ContentErrorException(ContentErrorCode.INVALID_CONTENT_ID));
 
         List<Comment> comments = commentRepository.findCommentsByContentId(contentId);
+
         List<CommentDTO> response = new ArrayList<>();
 
         for (Comment comment : comments) {
@@ -39,19 +46,26 @@ public class CommentService {
             response.add(dto);
         }
 
-        return response;
+        return SuccessResponse.from(response);
     }
 
     /**
      * 댓글 생성
      * @param req
      */
-    public CreateCommentRes createComment(CreateCommentReq req) {
-        User user = userRepository.findById(req.getUserId())
-                .orElseThrow(() -> new EntityNotFoundException("User does not exist"));
+    public SuccessResponse createComment(CreateCommentReq req, CustomUserDetails securityUser) {
+        User user = userRepository.findById(securityUser.getUserId()).orElse(null);
 
         Content content = contentRepository.findById(req.getContentId())
-                .orElseThrow(() -> new EntityNotFoundException("Content does not exist"));
+                .orElseThrow(() -> new ContentErrorException(ContentErrorCode.INVALID_CONTENT_ID));
+
+        if (req.getParentId() != null) {
+            Comment parentComment = commentRepository.findById(req.getParentId())
+                    .orElseThrow(() -> new CommentErrorException(CommentErrorCode.INVALID_PARENT_COMMENT));
+            if (parentComment.getDeleteAt() != null) {
+                throw new CommentErrorException(CommentErrorCode.INVALID_PARENT_COMMENT);
+            }
+        }
 
         Comment newComment = Comment.builder()
                 .content(content)
@@ -62,7 +76,7 @@ public class CommentService {
 
         commentRepository.save(newComment);
 
-        return CreateCommentRes.from(req, newComment);
+        return SuccessResponse.from(CreateCommentRes.from(req, newComment));
     }
 
     /**
@@ -70,25 +84,37 @@ public class CommentService {
      * @param commentId
      */
     @Transactional
-    public DeleteCommentRes deleteComment(Long commentId) {
+    public SuccessResponse deleteComment(Long commentId, CustomUserDetails securityUser) {
         Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new EntityNotFoundException("Comment does not exist"));
+                .orElseThrow(() -> new CommentErrorException(CommentErrorCode.INVALID_COMMENT_ID));
 
-        comment.deleteComment();
+        if (comment.getDeleteAt() != null)
+            throw new CommentErrorException(CommentErrorCode.DELETED_COMMENT);
 
-        return DeleteCommentRes.of(commentId, comment.getDeleteAt());
+        if (securityUser.getRole().equals("ROLE_ADMIN") || comment.getUser().getId().equals(securityUser.getUserId()))
+            comment.deleteComment();
+        else
+            throw new CommentErrorException(CommentErrorCode.REQUEST_USER_DATA_OWNER_MISMATCH);
+
+        return SuccessResponse.from(DeleteCommentRes.of(commentId, comment.getDeleteAt()));
     }
 
     /**
      * 댓글 수정
      * @param req
      */
-    public UpdateCommentRes updateComment(UpdateCommentReq req) {
+    public SuccessResponse updateComment(UpdateCommentReq req, CustomUserDetails securityUser) {
         Comment comment = commentRepository.findById(req.getCommentId())
-                .orElseThrow(() -> new EntityNotFoundException("Comment does not exist"));
+                .orElseThrow(() -> new CommentErrorException(CommentErrorCode.INVALID_COMMENT_ID));
 
-        commentRepository.save(comment.updateComment(req.getContent()));
+        if (comment.getDeleteAt() != null)
+            throw new CommentErrorException(CommentErrorCode.DELETED_COMMENT);
 
-        return UpdateCommentRes.of(comment.getId(), comment.getUpdatedAt());
+        if (securityUser.getRole().equals("ROLE_ADMIN") || comment.getUser().getId().equals(securityUser.getUserId()))
+            commentRepository.save(comment.updateComment(req.getContent()));
+        else
+            throw new CommentErrorException(CommentErrorCode.REQUEST_USER_DATA_OWNER_MISMATCH);
+
+        return SuccessResponse.from(UpdateCommentRes.of(comment.getId(), comment.getUpdatedAt()));
     }
 }
